refactor(nvme): shared vfio-pci bring-up helpers
Extract the invariant portion of the vfio-pci to NVMe-ready sequence
into nvme_controller_vfio_pci.h so nvme_controller_open_vfio and
nvme_controller_open_dmamem_vfio call the same code. The two
backends differ only in how they get a vfio device fd (type1 group
vs iommufd cdev bind), how DMA is mapped (VFIO_IOMMU_MAP_DMA vs
IOMMU_IOAS_MAP_FILE), and which heap backs the admin queue (hostmem
vs dmamem); everything from device-fd through CSTS.RDY=1 was
line-for-line duplicated across the two headers.

The shared helpers are nvme_vfio_pci_bus_master_enable (previously
a copy-paste in each backend), nvme_vfio_pci_acquire_bar0 (BM
enable + BAR0 region info + mmap + ctrlr func.bars population),
nvme_controller_reset_via_bar0 (CAP read + timeout_ms + CC disable
+ wait !RDY), and nvme_controller_enable_via_bar0 (CC.EN=1 with
the standard settings + wait RDY).
